Ajout admin Activity & Package
Ajout des Models et controlleurs de Activities et Packages:
- Ajout
- Modification
- Suppression

Ajout des liens de navigation dans la top bar

+++ routes/web.php
@@ -1,7 +1,9 @@
 <?php
 
+use App\Http\Controllers\ActivityController;
 use App\Http\Controllers\AdminController;
 use App\Http\Controllers\ArticleController;
+use App\Http\Controllers\PackageController;
 use App\Http\Controllers\UserController;
 use App\Models\Article;
 use Illuminate\Support\Facades\Route;
@@ -35,12 +37,29 @@ Route::get('/account/{id?}', [UserController::class, 'account'])->middleware(['a
 Route::match(['get', 'post'], '/account/modify/{id?}', [UserController::class, 'modify'])->middleware(['auth'])->name('user.modify');
 
 //Articles
-Route::get('admin/articles/', [AdminController::class, 'showArticles'])->name('admin.articles');
-Route::get('admin/articles/add', [ArticleController::class, 'add'])->name('admin.articles.add');
-Route::post('admin/articles/add/submit', [ArticleController::class, 'addSubmit'])->name('admin.articles.add.submit');
-Route::get('admin/articles/edit/{id}', [ArticleController::class, 'edit'])->name('admin.articles.edit');
-Route::post('admin/articles/edit/submit/{id}', [ArticleController::class, 'editSubmit'])->name('admin.articles.edit.submit');
-Route::delete('/admin/articles/remove/{id}', [ArticleController::class, 'remove'])->name('admin.articles.remove');
+Route::get('admin/articles/', [AdminController::class, 'showArticles'])->middleware(['auth'])->name('admin.articles');
+Route::get('admin/articles/add', [ArticleController::class, 'add'])->middleware(['auth'])->name('admin.articles.add');
+Route::post('admin/articles/add/submit', [ArticleController::class, 'addSubmit'])->middleware(['auth'])->name('admin.articles.add.submit');
+Route::get('admin/articles/edit/{id}', [ArticleController::class, 'edit'])->middleware(['auth'])->name('admin.articles.edit');
+Route::post('admin/articles/edit/submit/{id}', [ArticleController::class, 'editSubmit'])->middleware(['auth'])->name('admin.articles.edit.submit');
+Route::delete('/admin/articles/remove/{id}', [ArticleController::class, 'remove'])->middleware(['auth'])->name('admin.articles.remove');
+
+//Packages
+Route::get('admin/packages/', [AdminController::class, 'showPackages'])->middleware(['auth'])->name('admin.packages');
+Route::get('admin/packages/add', [PackageController::class, 'add'])->middleware(['auth'])->name('admin.packages.add');
+Route::post('admin/packages/add/submit', [PackageController::class, 'addSubmit'])->middleware(['auth'])->name('admin.packages.add.submit');
+Route::get('admin/packages/edit/{id}', [PackageController::class, 'edit'])->middleware(['auth'])->name('admin.packages.edit');
+Route::post('admin/packages/edit/submit/{id}', [PackageController::class, 'editSubmit'])->middleware(['auth'])->name('admin.packages.edit.submit');
+Route::delete('/admin/packages/remove/{id}', [PackageController::class, 'remove'])->middleware(['auth'])->name('admin.packages.remove');
+
+//Activities
+Route::get('admin/activities/', [AdminController::class, 'showActivities'])->middleware(['auth'])->name('admin.activities');
+Route::get('admin/activities/add', [ActivityController::class, 'add'])->middleware(['auth'])->name('admin.activities.add');
+Route::post('admin/activities/add/submit', [ActivityController::class, 'addSubmit'])->middleware(['auth'])->name('admin.activities.add.submit');
+Route::get('admin/activities/edit/{id}', [ActivityController::class, 'edit'])->middleware(['auth'])->name('admin.activities.edit');
+Route::post('admin/activities/edit/submit/{id}', [ActivityController::class, 'editSubmit'])->middleware(['auth'])->name('admin.activities.edit.submit');
+Route::delete('/admin/activities/remove/{id}', [ActivityController::class, 'remove'])->middleware(['auth'])->name('admin.activities.remove');
+
 
 /*
 |--------------------------------------------------------------------------
